requirements.txt
Django==2.0
djangorestframework==3.7.7
gunicorn==19.7.1
dj-database-url==0.4.2